Move collection.check_artifact_file into utils/
